fixed #119
implementation of the BandwidthTester-Tool.

The DownloadFile-ServerAction is used to download a given file from the
callserver. The bandwidth is determined by measuring the lapsed download
time, and dividing it from the size of the downloaded file.

A login prompt is preventing an unauthorized access.

needed parameters are:
args[0] - callserverUrl
args[1] - domain
args[2] - file (which have to exists on the server)