#124 added new class SqlConnectionException
this class is used since the Conncetion interfaces enforces the delete /
insert / update metaobjects method to throw an ConnectioException. This
class is used to indicate that during an server operation a SQL has
occured. In case such a an exception is thrown, we use a custom error
dialog that is shown to the user