fixed #129
QuerySearchResultsActionDialog now supports also a collection
of actions. if the collection contains only one action, the
panel looks and works exactly like before.
if the collection contains more than one action, a combobox
is displayed, that give the use the choice of the action that
have to be executed.

The functionality of the dialog has also been moved to
QuerySearchResultsActionPanel, which the Dialog includes.
This allow the use of the panel outside of a dialog.