fixed #151
The code in the process method of the worker is now only executed
if the worker arent been canceled.

The worker variable isnt set to null anymore when the worker is done,
because another (a later one) worker could exist.

Lookup for shared handlers is now done only once => loading speadup